% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPrediction.R
\name{krr_pred}
\alias{krr_pred}
\title{Predict new krr outcomes.}
\usage{
krr_pred(krr_mod, xnew, ynew = NA)
}
\arguments{
\item{krr_mod}{A model produced by the \code{krr} function}

\item{xnew}{New quantitative explanatory variables to predict (vector,
matrix, or data.frame).}

\item{ynew}{Optional new response variables, helps compute MSE.}
}
\value{
Predicted values and MSE.
}
\description{
A function for making predictions based on a krr (kernel ridge regression)
model.
}

