% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krr-package.R
\docType{package}
\name{krr-package}
\alias{krr}
\alias{krr-package}
\title{krr: Kernel Ridge Regression for Phonological Analogy}
\description{
Fit and apply kernel ridge regression models using precomputed phonological
distances. Key functions: \code{\link[=train_krr]{train_krr()}}, \code{\link[=predict_krr]{predict_krr()}}, \code{\link[=check_krr_inputs]{check_krr_inputs()}}.
}
\details{
See \code{vignette("lakok", package = "krr")} for a worked example.
}
\author{
\strong{Maintainer}: Peter Racz \email{petermartonracz@gmail.com}

}
\keyword{internal}
